requirements.txt
Pyrogram
asyncio
async-timeout
psutil
py-cpuinfo
uptime